requirements-api.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pandas==2.1.4
numpy==1.26.3
pyspark==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.34
sqlalchemy==2.0.25
pyarrow==15.0.0
delta-spark==3.0.0
kafka-python==2.0.2